Organized certain logic into base class.
